moved admin functions and views from pages to admin
