+++ app/controllers/pages_controller.rb
@@ -6,26 +6,6 @@ class PagesController < ApplicationController
     @name = current_user.first_name if current_user
   end
 
-  # Admin Only
-  def users
-    @buyers = User.buyers.all
-    @brokers = User.brokers.where(verified: true)
-    @users = User.all
-  end
-
-  def pending
-    @pending = User.brokers.where(verified: false)
-  end
-
-  def edit_pending
-    @user = User.find(params[:id])
-  end
-
-  def approve_pending
-    @user = User.find(params[:id])
-    redirect_to users_pending_path if @user.update(user_params)
-  end
-
   # Brokers and Buyers Only
   def portfolio
     case @user_type
