+++ app/helpers/admin_helper.rb
@@ -1,0 +1,2 @@
+module AdminHelper
+end
