+++ app/views/admin/pending.html.erb
@@ -6,13 +6,6 @@
   <table class="table table-striped table-bordered table-hover">
     <thead class="thead-dark">
       <tr>
-        <td><%= user.first_name %> <%= user.last_name %></td>
-        <td><%= user.email %></td>
-        <td>
-          <%= form_with model: user, url: users_approve_pending_path(user), method: :patch, local: true do |f| %>
-            <%= f.check_box :verified, onclick: "this.form.submit();" %> <br />
-          <% end %>
-        </td>
         <th scope="col">Name:</th>
         <th scope="col">Email:</th>
         <th scope="col">Approval:</th>
@@ -24,8 +17,7 @@
           <td><%= user.first_name %> <%= user.last_name %></td>
           <td><%= user.email %></td>
           <td>
-            <%= form_with model: user, url: "pending/#{user.id}", method: :patch, local: true do |f| %>
-              <%= f.hidden_field :id, :value => user.id %>
+            <%= form_with model: user, url: users_approve_pending_path(user), method: :patch, local: true do |f| %>
               <%= f.check_box :verified, onclick: "this.form.submit();" %> <br />
             <% end %>
           </td>
